+++ example/src/random-color-generator.ts
