--- example/webpack.config.js
@@ -1,6 +1,0 @@
-export const module = {
-  loaders: [{
-    test: /\.css$/,
-    loader: "style-loader!css-loader"
-  }]
-};
