+++ package.json
@@ -1,28 +1,27 @@
 {
   "name": "react-scrollable-feed",
-  "version": "1.3.2",
+  "version": "1.4.0",
   "description": "",
-  "author": "dizco",
-  "license": "MIT",
-  "repository": "dizco/react-scrollable-feed",
-  "main": "dist/index.js",
-  "module": "dist/index.es.js",
-  "jsnext:main": "dist/index.es.js",
-  "engines": {
-    "node": ">=14",
-    "npm": ">=5"
+  "type": "module",
+  "source": "src/index.tsx",
+  "main": "dist/index.cjs",
+  "module": "dist/index.module.mjs",
+  "unpkg": "dist/index.umd.js",
+  "types": "dist/index.d.ts",
+  "exports": {
+    "types": "./dist/index.d.ts",
+    "require": "./dist/index.cjs",
+    "default": "./dist/index.modern.mjs"
   },
   "scripts": {
-    "test": "cross-env CI=1 react-scripts-ts test --env=jsdom",
-    "test:watch": "react-scripts-ts test --env=jsdom",
-    "build": "rollup -c",
-    "start": "rollup -c -w",
-    "start:silent": "npm start -- --silent",
-    "prepare": "npm run build",
-    "predeploy": "cd example && npm install && npm run build",
-    "deploy": "gh-pages -d example/build",
-    "cypress:open": "cross-env CYPRESS_CRASH_REPORTS=0 cypress open",
-    "cypress:run": "cross-env CYPRESS_CRASH_REPORTS=0 cypress run"
+    "dev": "microbundle watch",
+    "build": "microbundle"
+  },
+  "author": "dizco",
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/dizco/react-scrollable-feed.git"
   },
   "peerDependencies": {
     "prop-types": "^15.7.2",
@@ -30,29 +29,7 @@
     "react-dom": "^15.0.0 || ^16.0.0 || ^17.0.0 || ^18.0.0"
   },
   "devDependencies": {
-    "@rollup/plugin-babel": "^6.0.4",
-    "@rollup/plugin-commonjs": "^25.0.7",
-    "@rollup/plugin-node-resolve": "^15.2.3",
-    "@rollup/plugin-typescript": "^11.1.5",
-    "@rollup/plugin-url": "^8.0.2",
-    "@testing-library/jest-dom": "^6.1.6",
-    "@testing-library/react": "^14.1.2",
-    "@types/jest": "^29.5.11",
     "@types/react": "^18.2.46",
-    "@types/react-dom": "^18.2.18",
-    "babel-core": "^6.26.3",
-    "babel-runtime": "^6.26.0",
-    "cross-env": "^7.0.3",
-    "cypress": "^13.6.2",
-    "gh-pages": "^6.1.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "rollup": "^4.9.2",
-    "rollup-plugin-peer-deps-external": "^2.2.4",
-    "rollup-plugin-postcss-modules": "^2.1.1",
-    "typescript": "^5.3.3"
-  },
-  "files": [
-    "dist"
-  ]
+    "microbundle": "^0.15.1"
+  }
 }
