--- rollup.config.js
@@ -1,39 +1,0 @@
-import typescript from '@rollup/plugin-typescript'
-import commonjs from '@rollup/plugin-commonjs'
-import external from 'rollup-plugin-peer-deps-external'
-import postcss from 'rollup-plugin-postcss-modules'
-import resolve from '@rollup/plugin-node-resolve'
-import url from '@rollup/plugin-url'
-
-import pkg from './package.json'
-
-export default {
-  input: 'src/index.tsx',
-  output: [
-    {
-      file: pkg.main,
-      format: 'cjs',
-      exports: 'named',
-      sourcemap: true
-    },
-    {
-      file: pkg.module,
-      format: 'es',
-      exports: 'named',
-      sourcemap: true
-    }
-  ],
-  plugins: [
-    external(),
-    postcss({
-      modules: true,
-      writeDefinitions: true
-    }),
-    url(),
-    resolve(),
-    typescript({
-      rollupCommonJSResolveHack: true
-    }),
-    commonjs()
-  ]
-}
